// packages/core/src/common/types.ts
export interface ChartRenderOptions {
  type: string;
  xAxisCol: string;
  yAxisCols: string[];
  numericCols: string[];
  sortBy?: string;
  limitRows?: number;
  dateFormat?: string;
  useLogScale?: boolean;
  showGridX?: boolean;
  showGridY?: boolean;
  showDataLabels?: boolean;
  showLabels?: boolean;
  chartTitle?: string;
  legendPosition?: string;
  horizontalBars?: boolean;
  curveTension?: number;
  lineStyle?: string;
  pointStyle?: string;
  blurEffect?: boolean;
  hiddenSlices?: Set<string>;
  selectedPieValueCol?: string;
  seriesColors?: Map<string, string>;
  sliceColors?: Map<string, string>;
  textColor?: string;
}